Branch is updated with new salary data type double

+++ SaiManoj-Spring-REST-Mysql-App/src/main/java/com/server/entity/Employee.java
@@ -19,7 +19,7 @@ public class Employee {
 	private String name;
 	
 	@Min(20000)
-	private float salary;
+	private double salary;
 	
 	public Employee() {
 		super();
